Addresses #132.

(Easier to play around with this for me when it's a PR.)