@pablogsal 2020-06-07
I have addressed most of the feedback (there are still some items like removing the macro or other suggestions).

I had to add another stack for the end positions because we need to recover for every pair of operator/factor the last non-whitespace token. For instance here:

```
a + b + (c      ) + d
```

the expression for `c` will end in `c` but the BinOP should point to the right parenthesis. That information can only be easily obtained when constructing the pair, but we need to use it later when constructing the BinOP. The most straightforward version that I could think of is another stack but maybe you have some better idea on how to move that information around more efficiently.