@gvanrossum 2020-06-07
Another idea: Make _ PyPegen_get_binary_operator into a function that calls _Py_BinOp -- then we can easily extend this scheme to include comparisons and Boolean AND/OR.

I think there's an extension to unary operators possible too, but I haven't figured it out yet.