@pablogsal 2020-06-07
> Another idea: Make _ PyPegen_get_binary_operator into a function that calls _Py_BinOp -- then we can easily extend this scheme to include comparisons and Boolean AND/OR.

Given that this PR is already big enough, I would prefer to do that in a future PR if you don't mind to reduce the moving parts in this one.