@pablogsal 2020-06-07
> Let me work on this a bit. I feel that the values and operators arrays can be merged, since an invariant is `stack_pos+1 == output_pos`, making the signature of the reduce function simpler.

Cool, feel free to push to the branch :)