@gvanrossum 2020-06-07
Hm, I have to think about this more and right now I've got to do other stuff. Later tonight I'll get back to it. For now I just pushed a tiny cleanup commit.