@gvanrossum 2020-06-07
Okay, it was trickier than I thought, but I think I have a version where everything's in the same struct, and eliminated `stack_pos` by replacing it with `output_pos - 1`.

The really subtle thing that tripped me up: when we write `values[output_pos++] = <new struct>` this overwrites the `oper` field (with NULL I suppose), but that's okay because the `.oper` field always trails one behind. This means that the first `.oper` field written goes next to the *initial* factor -- where the `.original_pair` field is not used.