@pablogsal 2020-06-07
> Okay, it was trickier than I thought, but I think I have a version where everything's in the same struct, and eliminated `stack_pos` by replacing it with `output_pos - 1`.

Thanks! I have repeated the benchmarks and I get the same as before (23% faster for simple inputs like `[1,2,3,4,5,6,7,8,9,10]` and 8% faster for xxl.py), so I would say this has been a success :)

This PR
-----------
```
❯ ./python -m pyperf timeit 'import _peg_parser' '_peg_parser.parse_string("2*3 + 4*5*6\n12 + (2 * 3 * 4 * 5 + 6 + 7 * 8)\n1 + 2 + 4 + 5 + 6 + 7 + 8 + 9 + 10 + ((((((11 * 12 * 13 * 14 * 15 + 16 * 17 + 18 * 19 * 20))))))", ast=False)'
.....................
Mean +- std dev: 25.2 us +- 0.1 us

❯ ./python -m pyperf timeit 'import _peg_parser' '_peg_parser.parse_string("[1,2,3,4,5,6,7,8,9,0]", ast=False)'
.....................
Mean +- std dev: 10.5 us +- 0.0 us
```

Master
----------
```
❯ ./python -m pyperf timeit 'import _peg_parser' '_peg_parser.parse_string("2*3 + 4*5*6\n12 + (2 * 3 * 4 * 5 + 6 + 7 * 8)\n1 + 2 + 4 + 5 + 6 + 7 + 8 + 9 + 10 + ((((((11 * 12 * 13 * 14 * 15 + 16 * 17 + 18 * 19 * 20))))))", ast=False)'
.....................
Mean +- std dev: 27.0 us +- 0.1 us

❯ ./python -m pyperf timeit 'import _peg_parser' '_peg_parser.parse_string("[1,2,3,4,5,6,7,8,9,0]", ast=False)'
.....................
Mean +- std dev: 13.7 us +- 0.1 us
```

Should I open a PR against CPython master?