@pablogsal 2020-06-07
For `time_stdlib` i get 12% speedup!

 This PR
-----------
```
./venv/bin/python scripts/test_parse_directory.py \
        -d ../../Lib \
        --short \
        --exclude "*/bad*" \
        --exclude "*/lib2to3/tests/data/*"
Checked 1,688 files, 775,043 lines, 28,549,953 bytes in 3.192 seconds.
That's 242,771 lines/sec, or 8,942,875 bytes/sec.
Memory stats:
  rss         :         33 MiB
  vms         :         38 MiB
  maxrss      :         34 MiB
```

Master
---------
```
./venv/bin/python scripts/test_parse_directory.py \
        -d ../../Lib \
        --short \
        --exclude "*/bad*" \
        --exclude "*/lib2to3/tests/data/*"
Checked 1,688 files, 775,059 lines, 28,550,369 bytes in 3.608 seconds.
That's 214,831 lines/sec, or 7,913,609 bytes/sec.
Memory stats:
  rss         :         30 MiB
  vms         :         35 MiB
  maxrss      :         35 MiB
```