@gvanrossum 2020-06-07
Thanks for confirming the benchmark results!

> Should I open a PR against CPython master?

Please do.

I have a few things I'd like to do but they can wait and I don't know when I'll have time for them, and I think it's safe to proceed without these:

- Rename values to stack and output_pos to stack_top. (I can do this now.)
- Replace the struct assignments to values[output_pos++] with individual field assignments (since this unnecessarily overwrites the oper field).
- Add some ASCII art explaining what's going on and what's on the stack at various points. (This takes careful thought and some paper and pencil work first.)
- Reason through whether the `oper` field can be replaced by accessing the original_pair->oper field of the next item. (This also requires paper and pencil work.)