@gvanrossum 2020-06-07
Took care of first two bullets. ASCII art put off till tonight. Final bullet is *wrong* -- that equivalence does not apply.