@pablogsal 2020-06-07
> Took care of first two bullets. ASCII art put off till tonight. 

Fantastic! I have opened an issue and a PR: https://bugs.python.org/issue40902