@gvanrossum 2020-06-07
Thanks. Please see if you still understand the last two functions in the file after I touched them; in particular I'm not sure if I've made all the stack operations clearer or less clear...