@pablogsal 2020-06-07
> Please see if you still understand the last two functions in the file after I touched them; in particular I'm not sure if I've made all the stack operations clearer or less clear...

Will check as soon as I finish dinner. 

Btw, I have pushed a commit to raise in the c generator if the associativity of the operators is not 'left' as currently we only support left-associative operators (also added a TODO item for that).