@gvanrossum 2020-06-07
> Btw, I have pushed a commit to raise in the c generator if the associativity of the operators is not 'left' as currently we only support left-associative operators (also added a TODO item for that).

I don't see it yet in the branch.