@gvanrossum 2020-06-08
Now a CPython PR (https://github.com/python/cpython/pull/20696) and bpo issue (https://bugs.python.org/issue40902).